bot/requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
asgiref==3.5.2
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncprawcore==2.3.0
attrs==21.4.0
beautifulsoup4==4.11.1
blis==0.7.8
catalogue==2.0.8
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.3
colorama==0.4.5
comtypes==1.1.11
cryptography==37.0.2
cymem==2.0.6
discord.py==1.7.3
Django==4.0.5
django-secrets==1.2.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz
future==0.18.2
geojson==2.5.0
h11==0.13.0
idna==3.3
Jinja2==3.1.2
joblib==1.1.0
keyboard==0.13.5
langcodes==3.3.0
MarkupSafe==2.1.1
mathparse==0.1.2
mongoengine==0.24.2
multidict==6.0.2
murmurhash==1.0.7
nltk==3.5
numpy==1.22.4
opencv-python==4.6.0.66
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pathy==0.6.2
Pint==0.19.2
praw==7.6.0
prawcore==2.3.0
preshed==3.0.6
PyAudio==0.2.12
pycparser==2.21
pydantic==1.9.1
pygame==2.1.2
pymongo==3.12.3
pyOpenSSL==22.0.0
pyowm==3.3.0
pyparsing==3.0.9
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.7.5
python-decouple==3.6
python-dotenv==0.20.0
pyttsx3==2.90
pytz==2022.1
pywin32==304
PyYAML==3.13
regex==2022.7.25
requests==2.28.0
schedule==1.1.0
selenium==4.2.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.9
spacy-loggers==1.0.3
SpeechRecognition==3.8.1
SQLAlchemy==1.2.19
sqlparse==0.4.2
srsly==2.4.4
TextToOwO==2.0.1
thinc==8.1.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typer==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
update-checker==0.18.0
urllib3==1.26.9
wasabi==0.10.1
webdriver-manager==3.7.0
websocket-client==1.3.3
wsproto==1.1.0
yarl==1.8.1
